+++ R/simva.R
@@ -9,6 +9,13 @@
 #' (default: `FALSE`).
 #' @param pambient `numeric(1)`, ambient pressure in kPa.
 #' @param pwater `numeric(1)`, water pressure in kPa.
+#' @param use_concentration_effect `logical(1)`, should concentration effect
+#' take into account (default: `FALSE`).
+#' @param tp_factor `numeric`, temperature/pressure factor.
+#' @param alveolar_minute_ventilation `numeric(1)`, alveolar minute ventilation
+#' in l/min.
+#' @param partial_pressures `logical(1)`, initial partial pressures
+#' settings to start with.
 #' @return `matrix`, with partial pressures for each simulation step.
 #' @export
 #'
@@ -74,79 +81,62 @@ sim_anaesthetic_uptake <- function(pinsp,
                                    conductances, capacitances,
                                    use_humidification = FALSE,
                                    pambient = 101.325,
-                                   pwater = 6.26
+                                   pwater = 6.26,
+                                   use_concentration_effect = FALSE,
+                                   tp_factor = stp_factor(),
+                                   alveolar_minute_ventilation =
+                                       conductances["lung"] / tp_factor,
+                                   partial_pressures =
+                                       c(pinsp = pinsp, lung = 0, vrg = 0,
+                                         mus = 0, fat = 0, cv = 0)
                                    ) {
+    nms <- c("pinsp", "lung", "vrg", "mus", "fat", "cv")
+
     n <- ceiling(total_time / delta_time)
 
     results <- matrix(
-        0.0, nrow = n, ncol = 7L,
-        dimnames = list(
-            c(),
-            c("time", "pinsp", "lung", "vrg", "mus", "fat", "cv")
-        )
+        NA_real_, nrow = n, ncol = 7L, dimnames = list(c(), c("time", nms))
     )
+
+    if (length(partial_pressures) != length(nms) ||
+            any(names(partial_pressures) != nms))
+        stop(
+            "'partial_pressures' has to be of length ", length(nms),
+            " and has to have the following names: ",
+            paste0(nms, collapse = ", ")
+        )
 
     if (isTRUE(use_humidification))
-        results[1L, "pinsp"] <- pinsp * (pambient / (pambient + pwater))
-    else
-        results[1L, "pinsp"] <- pinsp
+        partial_pressures["pinsp"] <-
+            partial_pressures["pinsp"] * (pambient / (pambient + pwater))
 
-    results[1L, ] <- c(
-        delta_time,
-        .sim_step(
-            results[1L, -1],
-            conductances = conductances,
-            capacitances = capacitances,
-            delta_time = delta_time
-        )
-    )
+    for (i in seq_len(n)) {
+        dvdtpt <- (partial_pressures["pinsp"] - partial_pressures["lung"]) *
+            conductances["lung"]
 
-    for (i in seq(from = 2L, to = n, by = 1L)) {
-        results[i, ] <- c(
-            i * delta_time,
-            .sim_step(
-                results[i - 1L, -1],
-                conductances = conductances,
-                capacitances = capacitances,
-                delta_time = delta_time
-            )
+        if (isTRUE(use_concentration_effect))
+            conductances["lung"] <-
+                alveolar_minute_ventilation * tp_factor + dvdtpt / 100
+
+        dvdt <- c(
+            dvdt1 = 0,
+            dvdt2 = (partial_pressures["lung"] - partial_pressures["vrg"]) *
+                conductances["vrg"],
+            dvdt3 = (partial_pressures["lung"] - partial_pressures["mus"]) *
+                conductances["mus"],
+            dvdt4 = (partial_pressures["lung"] - partial_pressures["fat"]) *
+                conductances["fat"]
         )
+
+        dvdt["dvdt1"] <- dvdtpt - sum(dvdt)
+        partial_pressures[c("lung", "vrg", "mus", "fat")] <-
+            partial_pressures[c("lung", "vrg", "mus", "fat")] +
+            dvdt * delta_time / capacitances
+        partial_pressures["cv"] <- sum(
+            partial_pressures[c("vrg", "mus", "fat")] *
+                conductances[c("vrg", "mus", "fat")]
+        ) / sum(conductances[c("vrg", "mus", "fat")])
+        results[i, ] <- c(i * delta_time, partial_pressures)
     }
     results
-}
-
-#' Single simulation step
-#'
-#' @param partial_pressures `double(7)`, columns: pinsp, palv, part, pvrg,
-#' pmus, pfat, pcv.
-#'
-#' @noRd
-.sim_step <- function(partial_pressures,
-                      conductances, capacitances, delta_time) {
-    dvdtpt <- (partial_pressures["pinsp"] - partial_pressures["lung"]) *
-        conductances["lung"]
-
-    # if concentration effect
-    # AMV = 4
-    # conductances["lung"] <- AMV * stp_factor() + dvdtpt / 100
-
-    dvdt <- c(
-        dvdt1 = 0,
-        dvdt2 = (partial_pressures["lung"] - partial_pressures["vrg"]) *
-            conductances["vrg"],
-        dvdt3 = (partial_pressures["lung"] - partial_pressures["mus"]) *
-            conductances["mus"],
-        dvdt4 = (partial_pressures["lung"] - partial_pressures["fat"]) *
-            conductances["fat"]
-    )
-
-    dvdt["dvdt1"] <- dvdtpt - sum(dvdt)
-    partial_pressures[c("lung", "vrg", "mus", "fat")] <-
-        partial_pressures[c("lung", "vrg", "mus", "fat")] +
-        dvdt * delta_time / capacitances
-    partial_pressures["cv"] <- sum(
-        partial_pressures[c("vrg", "mus", "fat")] *
-            conductances[c("vrg", "mus", "fat")]
-    ) / sum(conductances[c("vrg", "mus", "fat")])
-    partial_pressures
 }